requirements.txt
gensim
tqdm
nlp2
opencc-python-reimplemented
bz2file